Domains: Remove unused TLD validation schemas
