+++ client/state/action-types.ts
@@ -155,9 +155,6 @@ export const DOMAIN_MANAGEMENT_CONTACT_DETAILS_CACHE_REQUEST_SUCCESS =
 	'DOMAIN_MANAGEMENT_CONTACT_DETAILS_CACHE_REQUEST_SUCCESS';
 export const DOMAIN_MANAGEMENT_CONTACT_DETAILS_CACHE_UPDATE =
 	'DOMAIN_MANAGEMENT_CONTACT_DETAILS_CACHE_UPDATE';
-export const DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_ADD = 'DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_ADD';
-export const DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_REQUEST =
-	'DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_REQUEST';
 export const DOMAIN_MANAGEMENT_WHOIS_RECEIVE = 'DOMAIN_MANAGEMENT_WHOIS_RECEIVE';
 export const DOMAIN_MANAGEMENT_WHOIS_REQUEST = 'DOMAIN_MANAGEMENT_WHOIS_REQUEST';
 export const DOMAIN_MANAGEMENT_WHOIS_REQUEST_FAILURE = 'DOMAIN_MANAGEMENT_WHOIS_REQUEST_FAILURE';
