--- client/state/data-layer/wpcom/domains/validation-schemas/index.js
@@ -1,63 +1,0 @@
-import { get, flatMap } from 'lodash';
-import { DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_REQUEST } from 'calypso/state/action-types';
-import { bumpStat, composeAnalytics, recordTracksEvent } from 'calypso/state/analytics/actions';
-import { registerHandlers } from 'calypso/state/data-layer/handler-registry';
-import { http } from 'calypso/state/data-layer/wpcom-http/actions';
-import { dispatchRequest } from 'calypso/state/data-layer/wpcom-http/utils';
-import { addValidationSchemas } from 'calypso/state/domains/management/validation-schemas/actions';
-
-/**
- * Convert an application level request action for domain contact information
- * validation schemas into an HTTP request actions for the data-layer
- *
- * @param 	{object}        action      the schemas request action
- * @param   {Array<string>} action.tlds the tlds to be fetched from the API
- * @returns {object}                    The HTTP action for the data
- */
-export const fetch = ( action ) =>
-	http(
-		{
-			apiVersion: '1',
-			method: 'GET',
-			path: `/domains/validation-schemas/${ get( action, 'tlds', [] ).join( ',' ) }`,
-		},
-		action
-	);
-
-/**
- * Pop a HTTP request result into (an action to put it into) the state
- *
- * @param   {object} action   Originating action (unused).
- * @param   {object} schemas  Request result shaped like { tld: schema }
- * @returns {object}          An action to add the schema(s) to the state
- */
-export const onSuccess = ( action, schemas ) => addValidationSchemas( schemas );
-
-/**
- * Create an error notice action when the request fails
- *
- * @param   {object}      action Action object
- * @param   {Array}  action.tlds Originating action with the original list of requested tlds
- * @param   {object}       error Error information (query path, error message etc).
- * @returns {[object]}             An array of mc and tracks analytics events, one each per tld
- */
-export const onError = ( { tlds }, error ) =>
-	composeAnalytics(
-		...flatMap( tlds, ( tld ) => [
-			bumpStat( 'form_validation_schema_exceptions', `load_${ tld }` ),
-			recordTracksEvent( 'calypso_domain_contact_validation_schema_load_failure', {
-				tld,
-				error: JSON.stringify( error ),
-			} ),
-		] )
-	);
-
-registerHandlers( 'state/data-layer/wpcom/domains/validation-schemas/index.js', {
-	[ DOMAIN_MANAGEMENT_VALIDATION_SCHEMAS_REQUEST ]: [
-		dispatchRequest( {
-			fetch,
-			onSuccess,
-			onError,
-		} ),
-	],
-} );
