+++ client/state/domains/management/test/reducer.js
@@ -13,7 +13,6 @@ describe( 'reducer', () => {
 			'isRequestingContactDetailsCache',
 			'isRequestingWhois',
 			'isSaving',
-			'validationSchemas',
 		] );
 	} );
