--- client/state/selectors/get-validation-schemas.js
@@ -1,13 +1,0 @@
-import { get } from 'lodash';
-
-import 'calypso/state/domains/init';
-
-/**
- * Returns an object with domain validation schemas keyed by tld
- *
- * @param  {object}  state Global state tree
- * @returns {object|null} Validation JSON Schemas by tld
- */
-export default function getValidationSchemas( state ) {
-	return get( state, 'domains.management.validationSchemas', {} );
-}
